dptx: MST: Added asserts.

+++ XilinxProcessorIPLib/drivers/dptx/src/xdptx.h
@@ -792,11 +792,11 @@ u32 XDptx_ClearPayloadVcIdTable(XDptx *InstancePtr);
 /* xdptx_mst.c: Multi-stream transport (MST) functions for issuing sideband
  * messages. */
 u32 XDptx_SendSbMsgRemoteDpcdWrite(XDptx *InstancePtr, u8 LinkCountTotal,
-	u8 *RelativeAddress, u32 DpcdAddress, u8 BytesToWrite, u8 *WriteData);
+	u8 *RelativeAddress, u32 DpcdAddress, u32 BytesToWrite, u8 *WriteData);
 u32 XDptx_SendSbMsgRemoteDpcdRead(XDptx *InstancePtr, u8 LinkCountTotal,
-	u8 *RelativeAddress, u32 DpcdAddress, u8 BytesToRead, u8 *ReadData);
+	u8 *RelativeAddress, u32 DpcdAddress, u32 BytesToRead, u8 *ReadData);
 u32 XDptx_SendSbMsgRemoteIicRead(XDptx *InstancePtr, u8 LinkCountTotal,
-	u8 *RelativeAddress, u32 IicDeviceId, u8 BytesToRead, u8 *ReadData);
+	u8 *RelativeAddress, u8 IicDeviceId, u8 BytesToRead, u8 *ReadData);
 u32 XDptx_SendSbMsgLinkAddress(XDptx *InstancePtr, u8 LinkCountTotal,
 	u8 *RelativeAddress, XDptx_SbMsgLinkAddressReplyDeviceInfo *DeviceInfo);
 u32 XDptx_SendSbMsgEnumPathResources(XDptx *InstancePtr, u8 LinkCountTotal,
